almost fixed everything

+++ src/main/java/Servlets/DeletePostServlet.java
@@ -18,7 +18,15 @@ public class DeletePostServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         PostDAO postDAO = (SqlPostDAO) req.getServletContext().getAttribute("posts");
-        int post_id = Integer.parseInt(req.getParameter("adminText"));
+        String post_id_string = req.getParameter("adminText");
+
+        if(!isNumber(post_id_string)){
+            RequestDispatcher dispatcher = req.getRequestDispatcher("/adminPanel.jsp");
+            dispatcher.forward(req, resp);
+            return;
+        }
+
+        int post_id = Integer.parseInt(post_id_string);
 
         try {
             postDAO.removePost(post_id);
@@ -33,5 +41,14 @@ public class DeletePostServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws  ServletException, IOException{
         super.doGet(req, resp);
+    }
+
+    public static boolean isNumber(String str) {
+        try {
+            Integer.parseInt(str);
+        } catch (NumberFormatException e) {
+            return false;
+        }
+        return true;
     }
 }
